Add Fusion Coils (Sandbox 0.23.0), (Radial 0.7.0)
Added new weapons:
- [143] Kinetic Coil (Tier 1)
- [144] Hardlight Coil (Tier 2)
- [145] Plasma Coil (Tier 3)
- [146] Shock Coil (Tier 4)

We've added Fusion Coils to the sandbox. The Fusion Coil variants can be found under the "Equipment" menu instead of the "Weapons" menu, but purchasing one of them still requires a replacement of the primary or secondary weapon.
The Coils all cost 1 point, but are unlocked at different levels. This is because coild like the Plasma Coil and Shock Coil are better at dealing with vehicles, so they are unlocked later. This gives a small progression to obtaining the coils, so players would try to use the entire sandbox more.