Add weapons: Poisoned Pulse, AR Delta, ONI BR (Sandbox 0.23.1), (Radial 0.7.1)
Added new weapons:
- [56] Poisoned Pulse Carbine (Tier 3)
- [57] Assault Rifle Delta (Tier 3)
- [58] ONI Battle Rifle (Tier 2)

The Poisoned Pulse Carbine (Pulse Carbine) is an interesting needler-firing weapon at tier 3 that requires two supercombines to kill. It keeps the player in a fight for longer, but since the needler shots are tracking, the player can play around with their movement and positioning in a more creative way than when using ordinary weapons.

The Assault Rifle Delta (MA40 Assault Rifle) is just a desparate attempt at creating another MA40 Assault Rifle variant for tier 3. Its gimmick is that it does more damage and you need to shoot in controlled bursts to land your shots.

The ONI Battle Rifle (BR75) is a worse variant of the default BR75, but placed at tier 2, and costs points. The thought process behind it is that players get to use a BR75 variant very early on in the game. The weapon can still 4-burst perfect kill a player, but it has a very short red-reticle range and a guaranteed bullet spread, so landing that 4-burst perfect is very rare in practice.
This weapon took the name of the [6] ONI Battle Rifle, since the term "ONI" has changed during the development of TSG Warzone to mean a low-mid tier item instead of a high-tier one.