The Ace, Doom of Reach, Sentinel Beam adj. (Sandbox 0.23.1), (Radial 0.7.1)
Adjusted weapons:
- [6] ONI Battle Rifle
  - Name: ONI Battle Rifle → **The Ace**
  - Tier: 5 → **6**
  - Ammo: 14+70 → **14+84**
- [29] ONI Turret
  - Weapon Type: M247 HMG + Striker Sidekick → **M247 HMG + Impact Commando**
  - Weapon Damage: 0.30 → **0.34**
- [30] Doom Of Reach
  - Weapon Type: M247 HMG + MA40 Longshot → **M247 HMG + Striker Sidekick**
  - Weapon Damage: 1.20 → **0.40**
- [43] Banish Of Balaho
  - Shot Velocity: 3000 → **2000**
- [111] Sentinel Beam
  - Tier: 4 → **3**

The ONI Battle Rifle (BR75) was renamed to "The Ace" to distance the "ONI" term from high-tier weapons, and to give the weapon a cooler name. The weapon's tier was increased from 5 → 6 as it was already competing with the [62] BR75 Breacher at tier 5, and often beating it.
To account for the tier increase, the ammo capacity has been increased by one magazine, giving the weapon a higher kill potential, and making it fit nicer as a tier 6 weapon.

The ONI Turret (M247 HMG) had an overhaul to the weapon type with it changing to the Impact Commando config. This makes the weapon have a consistent spread and much less headshot prioritization. Before, the ONI Turret was a bit too good for a tier 4 in the hands of players who knew how to abuse the headshot prioritization of the Striker Sidekick config, but now the weapon is more in line with the long-range turret that it was originally designed to be.
The damage was adjusted to suit the new weapon config and to align with the expected damage of a weapon in a M247 HMG in tier 4.

The Doom Of Reach (M247 HMG) had an overhaul to he weapon type by changing it to the Striker Sidekick. The earlier MA40 Longshot config resulted in a slow projectile, which required players to lead their shots with the tier 7 weapon. Such leading of shots proved to be an annoyance for such a high-tier weapon that the change in weapon type was made to ease the use of the weapon at longer ranges.
The damage was adjusted to suite the new weapon config and retain the Time To Kill at the expected rate of the tier 7 M247 HMG.

The Banish Of Balaho (Skewer) custom projectile shot velocity was decreased from 3000 → 2000 as the earlier velocity was so fast, the projectile's hitbox was starting to distort, stretch, or get misplaced so that some part of the projectile would hit the player who shot it, causing them to take damage.
The lowered projectile speed is still enough to serve its purpose without causing the issue of damaging the shooter in the process.

The Sentinel Beam was dropped from tier 4 → 3 due to playtest feedback indicating that it didn't perform to the expectations of a tier 4 weapon.