New vehicles: Husky, Cougar; Dragon overhaul (Sandbox 0.23.2), (Radial 0.7.2)
Added new vehicles:
- [27] Husky (Tier 6)
- [28] Cougar (Tier 2)
- [29] Recon Falcon (Tier 4)
Adjusted vehicles:
- [19] Dragon
  - Weapon Type: [136] Wasp Weapon + Pursuit Hydra → **[141] Sentinel Beam + Convergence Bulldog**
  - Tier: 9 → **8**
  - Made the vehicle shoot Shock Coils.
Added new weapons:
- [141] Sentinel Beam + Convergence Bulldog
  - Notes: Used for detecting vehicle honk, not intended for players.
Adjusted weapons:
- [136] Wasp Weapon + Pursuit Hydra
  - Weapon Type: Wasp Weapon + Pursuit Hydra → **Rockethog Weapon + MA40 Longshot**

The Husky (Mongoose) is the ultimate Gungoose that shoots three Wasp Rockets at a time. The rockets have no tracking, so we will have to see how difficult it really is to use and if it fits in tier 6 in the long run.

The Cougar (Warthog) is the Scout Warthog from Halo 5, with it having no turret and an overall faster speed than a normal Warthog. The faster speed is achieved by a speed boost applied to the vehicle on honk. The boost can only be activated every 1.70 seconds.

The Recon Falcon (Falcon) has no side turrets, but features a nose gun operable by the pilot. The weapon is intentionally difficult to use since the vehicle is so low at tier 4 where granting players an effective aerial vehicle so early in the game will lead to snowballing.

The Dragon (Falcon) has been overhauled to ditch the rocket nose gun in favor of a more unique Shock Coil projectile. This vehicle serves as a way to shut down vehicle plays by being able to consistently shock them while piloting an aircraft.
The tier of the Dragon was decreased from 9 → 8 to account for the overhaul effects.