Add 8 new vehicles (Sandbox 0.24.0), (Radial 0.7.4)
Added new vehicles:
- [30] Breech-Load Razorback (Tier 4)
- [31] Impulse Rockethog (Tier 8)
- [32] Chopper (Tier 4)
- [33] Hyperius Chopper (Tier 6)
- [34] Chieftain Chopper (Tier 6)
- [35] Escharum Chopper (Tier 8)
- [36] Kig-Yar Ghost (Tier 5)
- [37] Spec Ops Warthog (Tier 7)

The Breech-Load Razorback (Razorback) features two Kinetic Coils attached to the back which can be shot out by honking the horn, and subsequently also exploded via the horn if they are detached from the back.
It's a fun gimmick vehicle, but we're still placing it at tier for due to concerns with the unlimited supply of Kinetic Coils which could potentially be exploited, so we will monitor how players react to the vehicle before making more adjustments.

The Impulse Rockethog (Gauss Hog) is the unreleased Gauss Hog which is its own vehicle type of a Warthog with a Gauss Cannon. The reason we've not included it in TSG Warzone until now was due to concerns with the lack of a visible trail for the projectile shot from the Gauss Cannon.
We're including it now for testing to see how it fits the sandbox and if players are annoyed by the missing trail or not, and then evaluate whether the vehicle still fits in to our sandbox.
The name of the vehicle is weird cause there isn't a string for "Gauss". This is the higherst tier Rockethog variant due to it being a two-shot kill for aerial vehicles and a one-shot for infantry at any range. The shot is quite difficult to land though as it has very minimal aim magnetism so the cannon user needs to be very precise.

The Chopper has been added to the sandbox this late due to earlier concerns about the splatter ability of the vehicle being able to destroy high-health vehicles and even bosses too easily. We are bringing it into the sandbox for testing and will be monitoring if it's actually a problem or not in real gameplay.

The Hyperius Chopper (Chopper) shoots Rockethog rockets which fit as a gun for a Chopper variant as they feature exploding rounds. Due to them still being rockets that also have a side-effect of displacing vehicles and units due to the explosion, the vehicle is placed at tier 6.

The Chieftain Chopper is a native vehicle variant, and is a unique version of the Chopper with a different fire rate, firing sound, splatter damage type, and vehicle color, so it fits perfectly into the TSG Warzone sandbox.
This vehicle is generally an improvement to the normal Chopper in multiple ways, so it's placed at tier 6.

The Escharum Chopper (Chieftain Chopper) is an upgraded Chieftain Chopper that shoots Riven Mangler shots. This is the ultimate single-manned fast-moving ground vehicle that can quickly deal with infantry and even challenge large vehicles due to its high damage output from the splatter.
The only targets this vehicle can't counter is aerial, and for these reasons it's placed at tier 8.

The Kig-Yar Ghost (Ghost) shoots Pinpoint Needler shots. It's only effective against infantry as the needles do basically no damage against vehicles. It's still tiered at 5 above the regular Ghost due to it being better at killing infantry targets. Playtesting will be required to determine whether this vehicle is a unique enough Ghost variant or not.

The Spec Ops Warthog (Warthog) is supposed to be the ultimate sleeper Warthog upgrade. It features a chaingun that operates just like the normal Warthog's, but it's more deadly and accurate. The vehicle also features a boost on honk just like the [28] Cougar to give it something extra to warrant a tier 7 Warthog placement.