Moonlight Wasp, Dragon adjustment; Vehicle renames (Sandbox 0.24.0), (Radial 0.7.4)
Adjusted vehicles:
- [17] Moonlight Wasp
  - Weapon Damage: 2.00 → **1.00**
- [19] Dragon
  - Shot Velocity: 200 → **300**
- [27] Husky
  - Name: Husky → **Cougar**
- [28] Cougar
  - Name: Cougar → **Scout Warthog**

The Moonlight Wasp (Wasp) has seen a damage decrease due to it overpowering all other tier 9 vehicle offerings most of the time, and also being very strong against aerial vehicles. Dropping the damage from 2.00 → 1.00 will still allow infantry to be killed as fast as it has been able to, but it reduces the vehicle-killing effectiveness of it.

The Dragon (Falcon) custom projectile shot velocity was increased from 200 → 300 to make landing accurate hits on targets easier, but to also reduce instances of an issue where any Shock Coils created by the Dragon will explode if one of them explodes. Increasing the shot velocity leaves the Shock Coils in the air for a shorter time thus decreasing the chances of the issue occurring and disrupting gameplay.

The Husky (Mongoose) was renamed to "Cougar" as the "Husky" name was bad.

The Cougar (Warthog) was renamed to "Scout Warthog" since we didn't realize the string "Scout" was available beforehand. This fits the vehicle better as it's the same name used for the vehicle in Halo 5 Warzone.