Valkyrie ammo, Demon movement, A Go At velocity up (Sandbox 0.24.0)
Added new weapons:
- [141] Rockethog Weapon + Riven Mangler (Vehicle weapon)
- [142] Rockethog Weapon + Striker Sidekick (Vehicle weapon)

Adjusted weapons:
- [13] Valkyrie
  - Ammo: 4+4 → **4+6**
- [36] Demon
  - Movement Speed: 1.00 → **1.20**
- [37] Spike Of Thav 'Sebarim
  - Ammo: 10+10 → **10+30** (+50)
- [41] A Go At
  - Rocket Jump velocity increase and scalar inversion fix
- [141] Sentinel Beam + Convergence Bulldog
  - [141] → **[146]**
  - Weapon Type: Sentinel Beam + Convergence Bulldog → **Plasma Cannon (object)**
- [143] Kinetic Coil
  - [143] → **[147]**
- [144] Hardlight Coil
  - [144] → **[148]**
- [145] Plasma Coil
  - [145] → **[149]**
- [146] Shock Coil
  - [146] → **[150]**

The Valkyrie (S7 Sniper Rifle) ammo was increased from 4+4 → 4+6 to see if it starts getting abused in Warzone Ultra again.

The Demon (Infected Energy Sword) movement speed was increased at the request of player 'Precellence' stating that the weapon's counterpart "Prophet's Bane" gameplay was largely attributed to its increased movement speed, which our version didn't have.
The movement speed has been increased from 1.00 → 1.20 to make the walk speed faster. The sprint speed is also faster as a consequence, but sprinting takes the player off of camo. One thing we will be monitoring is whether the weapon becomes too effective at a faster movement speed, as the weapon doesn't lose the full camo when the player is moving uncrouched unlike in Halo 5.

The Spike Of Thav 'Sebarim (Cindershot) ammo capacity was increased from 10+10 → 10+30 due to the Spike Grenade projectiles being difficult to land. Giving players more ammo warrants the use of the tier 5 weapon even if they are bound to miss some shots.

The A Go At (Fuel Rod SPNKr) rocket jumping velocity was increased and an inversion issue in the script fixed. This allows for players to traverse larger distances with the rocket jumping mechanic, which is the main gimmick of the weapon, and was not seeing enough use beforehand.