Air vehicle tier decrease; Major vehicle changes (Sandbox 0.25.0), (Radial 0.8.0)
Adjusted vehicles:
- [1] Banshee
  - Tier: 7 → **6**
- [9] Wasp
  - Tier: 7 → **6**
- [11] Mid Banshee
  - Tier: 6 → **5**
- [16] Phantom Wasp
  - Tier: 8 → **7**
- [17] Moonlight Wasp
  - Weapon Damage: 1.00 → **0.80**
- [19] Dragon
  - Name: Dragon → **Paralyzing Falcon**
  - Shot Cadence: 2.00 → **3.50**
- [23] Carrier Falcon
  - Name: Carrier Faclon → **Dragon**
  - Tier: 3 → **9**
  - Brought back side turrets
  - Health Adjustment: 60% → **100%**
  - Weapon Type: None → **[141] Rockethog Weapon + Riven Mangler**
  - Shot Cadence: 1.00 → **0.03**
- [25] ONI Wasp
  - Point Cost: 7 → **6**
- [29] Recon Falcon
  - Weapon Damage: 0.70 → **1.20**
  - Weapon Type: [136] Rockethog Weapon + MA40 Longhsot → **Rockethog Weapon + BR75 Breacher**
  - Health Adjustment: 100% → **60%**
- [30] Breech-Load Razorback
  - Point Cost: 4 → **3**
- [36] Kig-Yar Ghost
  - Name: Kig-Yar Ghost → **Famished Ghost**
  - Weapon Type: [140] Rockethog Weapon + Pinpoint Needler → **[137] Rockethog Weapon + Scatterbound Heatwave**
  - Weapon Damage: 1.00 → **0.60**
  - Shot Cadence: 0.10 → **0.30**
- [37] Spec Ops Warthog
  - Weapon Type: [142] Rockethog Weapon + Striker Sidekick → **[132] Rockethog Weapon + Rapidfire Pulse Carbine**
  - Shot Cadence: 0.05 → **0.30**
  - Weapon Damage: 0.50 → **3.00**

Tiers of the aerial vehicles: Banshee, Wasp, Mid Banshee, Phantom Wasp & ONI Wasp have all beed decreased by 1 at the request of Warzone player 'Flamency'. We will see how these perform in the sandbox, but have a very short fuse on bringing them back up a tier if any overpowering issues arise from the change.

The Moonlight Wasp (Wasp) damage was decreased from 1.00 → 0.80 as a means to further decrease the damage it can deal to aerial vehicles in an effort to balance the tier 9 vehicle.

The Dragon (Falcon) was renamed to "Paralyzing Falcon" and the Carrier Falcon (Falcon) was rebranded to "Dragon", and multiple changes made to overhaul the vehicle to being a high-tier Falcon. The new [23] Dragon now shoots a stream of Riven Mangler shots out of the nose gun, leaning into a fire-breathing aesthetic of a "Dragon".
This change also removes the Carrier Falcon from the sandbox as it served no useful purpose in the sandbox than to just splatter unsuspecting players which was a playstyle we did not appreciate for a tier 3 vehicle.

The Recon Falcon (Falcon) has seen a change to the weapon type its using and a damage increase to make it easier to use. Previously the MA40 Longshot config it shot was too difficult to land shots with as it combined the bad articulation of the nose gun combined with the slow projectile speed of the MA40 Longshot.
The new weapon config of the BR75 Breacher has faster moving projectiles, and the increased damage helps with getting more damage on target. The vehicle is still at tier 4 for the lowest tier aerial shooting vehicle, and is intentionally difficult to use for that reason.
The health of the vehicle was adjusted to 60% to match that of the Carrier Falcon's, as the Recon Falcon now doubles as the Carrier Falcon, just also having a nose gun, but still no side turrets.

The Breech-Load Razorback (Razorback) was dropped from tier 4 → 3 due to feedback about it being too difficult to hit the launched Kinetic Coils on target as you have to usually turn your vehicle around to aim it at the target first, and even after firing, you aren't guaranteed to even be on target.

The Kig-Yar Ghost (Ghost) has been overhauled to be the "Famished Ghost" which shoots Scattebound Heatwave projectiles. The Kig-Yar Ghost was too underwhelming for a Ghost variant as it couldn't deal damage to vehicles, which we've learned we want to be a core aspect of the Ghosts since they natively deal plasma damage.
The Scatterbound Heatwave works in the favor of the anti-vehicle philosophy of the Ghost, and is also an interesting way to fight infantry.

The Spec Ops Warthog (Warthog) has been overhauled to shoot a Rapidfire Pulse Carbine and has had its shot cadence and weapon damaged adjusted accordingly to balance it for a tier 7 Warthog.
The previous Striker Sidekick-shooting Warthog turret proved to not show a visible enough tracer for the shots, and combined with no sound coming from the turret, it was too confusing for the driver and for enemies being attacked by the vehicle where they were being shot from.