2 Weapons, Marksman Sidekick ovh. Spike of Thav. (Sandbox 0.25.0), (Radial 0.8.0)
Removed weapons:
- [142] Rockethog Weapon + Striker Sidekick

Added weapons:
- [59] Burst-Fire Bandit (Tier 4)
- [60] SPNKr x (Tier 8)

Adjusted weapons:
- [1] Marksman Sidekick
  - Level: 2 → **4**
  - Weapon Type: Mk50 Sidekick + Impact Commando → **Mk50 Sidekick + MA40 Longshot**
  - Weapon Damage: 0.71 → **1.50**
  - Ammo: 40+60 → **25+75**
- [3] Hardlight Assault Rifle
  - Level: 3 → **2**
- [24] Focus Beam
  - Point Cost: 5 → **4**
- [36] Demon
  - Name: Demon → **The Demon**
  - Movement Speed: 1.20 → **1.25**
- [37] Spike of Thav 'Sebarim
  - Ammo: 10+30 → **10+20**
  - Shot Velocity: 400 → **300**
  - Added feature to track the shot projectile while zoomed.
- [38] Hunters Right Hand
  - Weapon Type: Fuel Rod SPNKr + Striker Sidekick → **Fuel Rod SPNKr + Impact Commando**
- [64] Calcine Disruptor
  - Point Cost: 6 → **5**
- [36] Demon
  - Name: Demon → The Demon
- [92] Bandit Evo
  - Level: 9 → **8**
- [112] Shock Rifle
  - Point Cost: 5 → **4**
- [136] Rockethog Weapon + MA40 Longshot → **Rockethog Weapon + BR75 Breacher**

The Burst-Fire Bandit (M392 Bandit) is a Bandit that shoots 3 shots at once like a BR75. The weapon has a lot of recoil and a bullet spread that maxes out right after the first shot is fired, so you cannot be accurate with it. That being said, the weapon is capable of a 2-burst kill, but you'll have to be very close to hit all shots on the body and head.

The SPNKr x (Fuel Rod SPNKr) is a cross between a Fuel Rod SPNKr and a M41 Tracker that has the benefits of both. It shoots tracking rockets, but the gimmick is that you must hit a player with a direct hit in order for a one-shot kill. You can even shoot at your own feet with the rocket and you won't die.
The weapon is still the highest Fuel Rod SPNKr due to its ease of use for both infantry and vehicular targets.

The Marksman Sidekick (Mk50 Sidekick) has received yet another overhaul, which returns the config of the weapon back to the MA40 Longshot. We never felt like this weapon filled its role of a long-range pistol variant that well and then we realized that lets just make it a better short–medium range sidekick instead of trying to make it a long-range weapon like we already have many other offerings of in the loadout weapon selection.
The main difference this time is that we aren't artificially decreasing the damage of the weapon as a balance measure for it to stay at level 2 since now it has been moved to level 4 and we can adjust its damage accordingly for that level of loadout weapon.
We've redesigned this weapon to be able to kill a player in one shot less than the Mk50 Sidekick, but it still has the MA40 Longshot config, which has a slow projectile speed, so effectiveness of the weapon lowers with range. We the weapon now fills a good role as a short–medium range sidekick pistol, which none of the other loadout weapons fill the role of.

The Hardlight Assault Rifle (MA40 Assault Rifle) and Bandit Evo have seen level decreases by 1 to suit in the small loadout weapon adjustment that also includes the Marksman Sidekick overhaul.
The Hardlight Assault Rifle was moved from level 3 → 2 to fill a role as an unlockable loadout weapon at 2, and we also hadn't seen players using the weapon too much at level 3.
The Bandit Evo decrease from level 9 → 8 is something we are testing out due to feedback from Warzone player 'Precellence' stating that players like to use the Bandit Evo and that it's being unlocked too late in the game at level 9, which is leaving out a core weapon of the Halo Infinite sandbox to be unused.
Based on our testing, the Bandit Evo is the ideal loadout weapon for both controller and mouse inputs, and we don't want to change its status as that, so we are making this change cautiously and will monitor how it feels at level 8.

The Focus Beam (Sentinel Beam) tier was decresed from 5 → 4 due to getting more experience with tier 5 weapons and realizing it wasn't living up to the standards of a tier 5 weapon.

The Demon (Energy Sword) has seen a name change to "The Demon" and following feedback from Warzone player 'Precellence' a movement increase from 1.20 → 1.25. We will monitor whether this speed increase makes the weapon too powerful as our variant of the weapon doesn't make players go out of camouflage unlike the "Prophet's Bane" from Halo 5, which the weapon is based off of.

The Spike Of Thav 'Sebarim (Cindershot) has seen a major improvement in it's consistency with the addition of projectile tracking to the aim position while zoomed. This projectile tracking/controlling works the same way as the default Cindershot where the projectile will be guided in the direction of the position where a raycast from your aim position hits, allowing you to control the trajectory of the projectile.
With the addition of the projectile controlling, the ammo amount has been decreased from 10+30 → 10+20 to compensate making it easier to hit targets. The projectile speed has also been adjusted from 400 → 300 to more closely match the default Cindershot projectile speed, making the custom weapon feel more like a part of the native sandbox.

The Hunters Right Hand (Fuel Rod SPNKr) has seen a config change from the Striker Sidekick → Impact Commando solely for distinguishing the weapon config from the SPNKr x, due to the crosshair size being larger with the Impact Commando config.

The Calcine Disruptor and Shock Rifle have seen a tier decrease of 1 following an initiative to make shock weapons more available. We will monitor the effects of this change, but are pretty confident that this is the right play. We have been cautious of placing shock weapons too low due to their high effectiveness at stopping vehicles, but have now gathered enough playtesting experience to be able to more accurately rank them in the sandbox.