Remove Assault Rifle Delta; Add Bassus Artifact (Sandbox 0.25.1), (Radial 0.8.1)
Removed weapons:
- [57] Assault Rifle Delta (Tier 3)

Added weapons:
- [57] Bassus Artifact (Tier 8)

The Assault Rifle Delta (MA40 Assault Rifle) has been removed due to it not being a unique enough of a weapon.
The Bassus Artifact (Gravity Hammer) is a unique Hammer variant that has the ability to zoom in and also to lunge at targets within red-reticle range. The weapon also features increased damage resistance, movement speed, jump height, melee impulse and melee recovery speed, making it the ultimate Gravity Hammer, rivaling that of the Rushdown Hammer.
The reason the weapon is still at tier 8 is cause it has a maximum of 10 swings of the Hammer.